+++ packages/contracts/src/geolocation/geolocation.services.pb.ts
@@ -8,7 +8,7 @@
 import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
 import { Observable } from "rxjs";
 import { TrackTransactionRequest } from "./geolocation.command.pb.js";
-import { GetNearbyTransactionsRequest } from "./geolocation.query.pb.js";
+import { GetMapZonesRequest, GetMapZonesResponse, GetNearbyTransactionsRequest } from "./geolocation.query.pb.js";
 import { GetNearbyTransactionsResponse, TrackTransactionResponse } from "./geolocation.responses.pb.js";
 
 export const protobufPackage = "geolocation";
@@ -19,6 +19,8 @@ export interface GeolocationServiceClient {
   trackTransaction(request: TrackTransactionRequest): Observable<TrackTransactionResponse>;
 
   getNearbyTransactions(request: GetNearbyTransactionsRequest): Observable<GetNearbyTransactionsResponse>;
+
+  getMapZones(request: GetMapZonesRequest): Observable<GetMapZonesResponse>;
 }
 
 export interface GeolocationServiceController {
@@ -29,11 +31,15 @@ export interface GeolocationServiceController {
   getNearbyTransactions(
     request: GetNearbyTransactionsRequest,
   ): Promise<GetNearbyTransactionsResponse> | Observable<GetNearbyTransactionsResponse> | GetNearbyTransactionsResponse;
+
+  getMapZones(
+    request: GetMapZonesRequest,
+  ): Promise<GetMapZonesResponse> | Observable<GetMapZonesResponse> | GetMapZonesResponse;
 }
 
 export function GeolocationServiceControllerMethods() {
   return function (constructor: Function) {
-    const grpcMethods: string[] = ["trackTransaction", "getNearbyTransactions"];
+    const grpcMethods: string[] = ["trackTransaction", "getNearbyTransactions", "getMapZones"];
     for (const method of grpcMethods) {
       const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
       GrpcMethod("GeolocationService", method)(constructor.prototype[method], method, descriptor);